refactor: shift types to correct dir
