+++ client/util/index.ts
@@ -1,6 +1,6 @@
 export * from "./constants";
+export * from "./types";
 export * from "./convertor";
 export * from "./mock";
-export * from "./types";
 export * from "./markdown-to-react-node";
 export * from "./date-utils";
