+++ client/util/types/models.ts
@@ -53,41 +53,3 @@ export type User = Util & {
   answers?: Answer[];
   questions?: Question[];
 };
-
-// -----------------------------------------------------------------------------
-// API Types
-// -----------------------------------------------------------------------------
-
-export type GetPaginatedQuestionsParam = {
-  page: number;
-  pageSize: number;
-};
-
-export type GetPaginatedQuestionRes = {
-  questions: Question[];
-  total: number;
-};
-
-export type GetSingleQuestionParam = {
-  title: string;
-  markdown: string;
-  level: string;
-  subject: string;
-};
-
-export type ApiErrorRes = {
-  status: string;
-  message: string;
-};
-
-export type RegisterUserParam = {
-  email: string;
-  password: string;
-};
-
-export type GetSingleUserRes = Util & {
-  email: string;
-  username: string;
-  answerIds: string[];
-  questionIds: string[];
-};
